Enhance StorageFilterInterface and StorageWrapper to be more complete.

+++ src/Config/StorageWrapper.php
@@ -9,6 +9,15 @@ namespace Drupal\config_filter\Config;
 
 use Drupal\Core\Config\StorageInterface;
 
+/**
+ * Class StorageWrapper.
+ *
+ * This class wraps another storage.
+ * It filters the arguments before passing them on to the storage for write
+ * operations and filters the result of read operations before returning them.
+ *
+ * @package Drupal\config_filter\Config
+ */
 class StorageWrapper implements StorageInterface {
 
   /**
@@ -49,11 +58,9 @@ class StorageWrapper implements StorageInterface {
    */
   public function read($name) {
     $data = $this->storage->read($name);
-
     foreach ($this->filters as $filter) {
       $data = $filter->filterRead($name, $data);
     }
-
     return $data;
   }
 
@@ -61,20 +68,11 @@ class StorageWrapper implements StorageInterface {
    * {@inheritdoc}
    */
   public function readMultiple(array $names) {
-    $dataList = $this->storage->readMultiple($names);
-    $result = [];
-
-    foreach ($names as $name) {
-      $data = isset($dataList[$name]) ? $dataList[$name] : NULL;
-      foreach ($this->filters as $filter) {
-        $data = $filter->filterRead($name, $data);
-      }
-      if ($data) {
-        $result[$name] = $data;
-      }
+    $data = $this->storage->readMultiple($names);
+    foreach ($this->filters as $filter) {
+      $data = $filter->filterReadMultiple($names, $data);
     }
-
-    return $result;
+    return $data;
   }
 
   /**
@@ -87,16 +85,20 @@ class StorageWrapper implements StorageInterface {
     if ($data) {
       return $this->storage->write($name, $data);
     }
-    return $this->storage->delete($name);
+    // The data was not written, but it is not an error.
+    return TRUE;
   }
 
   /**
    * {@inheritdoc}
    */
   public function delete($name) {
-    $success = $this->storage->delete($name);
+    $success = TRUE;
     foreach ($this->filters as $filter) {
       $success = $filter->filterDelete($name, $success);
+    }
+    if ($success) {
+      $success = $this->storage->delete($name);
     }
     return $success;
   }
@@ -105,7 +107,14 @@ class StorageWrapper implements StorageInterface {
    * {@inheritdoc}
    */
   public function rename($name, $new_name) {
-    return $this->storage->rename($name, $new_name);
+    $success = TRUE;
+    foreach ($this->filters as $filter) {
+      $success = $filter->filterRename($name, $new_name, $success);
+    }
+    if ($success) {
+      $success = $this->storage->rename($name, $new_name);
+    }
+    return $success;
   }
 
   /**
@@ -128,7 +137,7 @@ class StorageWrapper implements StorageInterface {
   public function listAll($prefix = '') {
     $data = $this->storage->listAll($prefix);
     foreach ($this->filters as $filter) {
-      $data = $filter->filterListAll($data, $prefix);
+      $data = $filter->filterListAll($prefix, $data);
     }
     return $data;
   }
@@ -137,28 +146,49 @@ class StorageWrapper implements StorageInterface {
    * {@inheritdoc}
    */
   public function deleteAll($prefix = '') {
-    return $this->storage->deleteAll($prefix);
+    $delete = TRUE;
+    foreach ($this->filters as $filter) {
+      $delete = $filter->filterDeleteAll($prefix, $delete);
+    }
+    if ($delete) {
+      $delete = $this->storage->deleteAll($prefix);
+    }
+    return $delete;
   }
 
   /**
    * {@inheritdoc}
    */
   public function createCollection($collection) {
-    return $this->storage->createCollection($collection);
+    $filters = [];
+    foreach ($this->filters as $filter) {
+      if ($filter = $filter->filterCreateCollection($collection)) {
+        $filters[] = $filter;
+      }
+    }
+    return new static($this->storage->createCollection($collection), $filters);
   }
 
   /**
    * {@inheritdoc}
    */
   public function getAllCollectionNames() {
-    return $this->storage->getAllCollectionNames();
+    $collections = $this->storage->getAllCollectionNames();
+    foreach ($this->filters as $filter) {
+      $collections = $filter->filterGetAllCollectionNames($collections);
+    }
+    return $collections;
   }
 
   /**
    * {@inheritdoc}
    */
   public function getCollectionName() {
-    return $this->storage->getCollectionName();
+    $collection = $this->storage->getCollectionName();
+    foreach ($this->filters as $filter) {
+      $collection = $filter->filterGetCollectionName($collection);
+    }
+    return $collection;
   }
 
 }
